refactor IXDTF implementation by concern
