--- ixdtf_test.go
@@ -1,1364 +1,0 @@
-package ixdtf_test
-
-import (
-	"errors"
-	"sort"
-	"strings"
-	"testing"
-	"time"
-
-	"github.com/8beeeaaat/ixdtf"
-)
-
-func TestFormat(t *testing.T) {
-	t.Parallel()
-	tokyo, paris, cet := getTestTimezones()
-
-	tests := []struct {
-		name    string
-		tm      time.Time
-		ext     *ixdtf.IXDTFExtensions
-		want    string
-		wantErr error
-	}{
-		{
-			name: "no extensions",
-			tm:   time.Date(2025, 1, 2, 3, 4, 5, 0, time.UTC),
-			ext:  ixdtf.NewIXDTFExtensions(nil),
-			want: "2025-01-02T03:04:05Z",
-		},
-		{
-			name: "use location as timezone",
-			tm:   time.Date(2025, 2, 3, 4, 5, 6, 0, tokyo),
-			ext:  ixdtf.NewIXDTFExtensions(nil),
-			want: "2025-02-03T04:05:06+09:00[Asia/Tokyo]",
-		},
-		{
-			name: "use specified timezone",
-			tm:   time.Date(2025, 1, 1, 0, 0, 0, 0, time.UTC),
-			ext: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{
-				Location: tokyo,
-			}),
-			want: "2025-01-01T00:00:00Z[Asia/Tokyo]",
-		},
-		{
-			// A critical time-zone flag without a zone cannot be honored;
-			// silently dropping the "!" would violate RFC 9557 Section 3.3.
-			name: "critical location flag without location errors",
-			tm:   time.Date(2025, 1, 1, 0, 0, 0, 0, time.UTC),
-			ext: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{
-				CriticalLocation: true,
-			}),
-			wantErr: ixdtf.ErrCriticalExtension,
-		},
-		{
-			name: "tags sorting and critical",
-			tm:   time.Date(2025, 3, 4, 5, 6, 7, 0, time.UTC),
-			ext: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{
-				Tags: map[string]string{
-					"b-tag": "2",
-					"a-tag": "1",
-				},
-				Critical: map[string]bool{
-					"b-tag": true,
-				},
-			}),
-			want: "2025-03-04T05:06:07Z[a-tag=1][!b-tag=2]",
-		},
-		{
-			name: "timezone and tags",
-			tm:   time.Date(2025, 6, 7, 8, 9, 10, 0, cet),
-			ext: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{
-				Location: paris,
-				Tags: map[string]string{
-					"u-ca": "gregory",
-				},
-				Critical: map[string]bool{
-					"u-ca": true,
-				},
-			}),
-			want: "2025-06-07T08:09:10+01:00[Europe/Paris][!u-ca=gregory]",
-		},
-		{
-			name: "private extension error",
-			tm:   time.Date(2025, 1, 1, 0, 0, 0, 0, time.UTC),
-			ext: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{
-				Tags: map[string]string{
-					"x-demo": "yes",
-				},
-			}),
-			wantErr: ixdtf.ErrPrivateExtension,
-		},
-		{
-			name: "invalid timezone",
-			tm:   time.Date(2025, 1, 1, 0, 0, 0, 0, time.UTC),
-			ext: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{
-				Location: time.FixedZone("No/SuchZone", 1*3600),
-			}),
-			wantErr: ixdtf.ErrInvalidTimezone,
-		},
-		{
-			name: "missing critical tag",
-			tm:   time.Date(2025, 1, 1, 0, 0, 0, 0, time.UTC),
-			ext: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{
-				Critical: map[string]bool{"u-ca": true},
-			}),
-			wantErr: ixdtf.ErrCriticalExtension,
-		},
-		{
-			name: "invalid critical calendar value",
-			tm:   time.Date(2025, 1, 1, 0, 0, 0, 0, time.UTC),
-			ext: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{
-				Tags:     map[string]string{"u-ca": "hoge"},
-				Critical: map[string]bool{"u-ca": true},
-			}),
-			wantErr: ixdtf.ErrInvalidTagCalendarIdentifier,
-		},
-		{
-			name: "empty timezone name should not add brackets",
-			tm:   time.Date(2025, 1, 1, 12, 0, 0, 0, time.UTC),
-			ext: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{
-				Location: time.FixedZone("", 0),
-			}),
-			want: "2025-01-01T12:00:00Z",
-		},
-		{
-			name: "nil extensions behaves like empty extensions",
-			tm:   time.Date(2025, 2, 3, 4, 5, 6, 0, tokyo),
-			ext:  nil,
-			want: "2025-02-03T04:05:06+09:00[Asia/Tokyo]",
-		},
-		{
-			// RFC 9557 Section 1.2: an offset time zone is serialized using
-			// the same numeric form as the RFC 3339 offset ("+09:00", not the
-			// Go zone-name convention "+0900").
-			name: "offset time zone serialized in RFC form",
-			tm:   time.Date(2025, 1, 1, 0, 0, 0, 0, time.FixedZone("+09:00", 9*3600)),
-			ext: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{
-				Location: time.FixedZone("+09:00", 9*3600),
-			}),
-			want: "2025-01-01T00:00:00+09:00[+09:00]",
-		},
-		{
-			name: "critical offset time zone keeps flag and RFC form",
-			tm:   time.Date(2025, 1, 1, 0, 0, 0, 0, time.FixedZone("+09:00", 9*3600)),
-			ext: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{
-				Location:         time.FixedZone("+09:00", 9*3600),
-				CriticalLocation: true,
-			}),
-			want: "2025-01-01T00:00:00+09:00[!+09:00]",
-		},
-		{
-			// The critical flag also applies to the timestamp's own named
-			// zone when ext.Location is unset — the same fallback used for
-			// non-critical output, so the "!" is not silently dropped.
-			name: "critical flag applies to fallback zone from timestamp",
-			tm:   time.Date(2025, 2, 3, 4, 5, 6, 0, tokyo),
-			ext: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{
-				CriticalLocation: true,
-			}),
-			want: "2025-02-03T04:05:06+09:00[!Asia/Tokyo]",
-		},
-	}
-
-	sort.Slice(tests, func(i, j int) bool { return tests[i].name < tests[j].name })
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			t.Parallel()
-			got, err := ixdtf.Format(tc.tm, tc.ext)
-
-			if tc.wantErr != nil {
-				if !errors.Is(err, tc.wantErr) {
-					t.Fatalf("expected error %v, got %v", tc.wantErr, err)
-				}
-				return
-			}
-
-			if err != nil {
-				t.Fatalf("unexpected error: %v", err)
-			}
-
-			if got != tc.want {
-				t.Fatalf("want %q got %q", tc.want, got)
-			}
-		})
-	}
-}
-
-// TestOffsetTimezoneRoundTrip verifies that offset time-zone annotations
-// survive a Parse -> Format -> Parse round trip in the RFC 9557 Section 1.2
-// serialization form, including the Section 1.2 example "+08:45[+08:45]" and
-// the critical variant.
-func TestOffsetTimezoneRoundTrip(t *testing.T) {
-	t.Parallel()
-
-	inputs := []string{
-		"2025-01-01T00:00:00+09:00[+09:00]",
-		"2022-07-08T00:14:07+08:45[+08:45]", // RFC 9557 Section 1.2 example
-		"2025-01-01T00:00:00+09:00[!+09:00]",
-	}
-
-	for _, input := range inputs {
-		t.Run(input, func(t *testing.T) {
-			t.Parallel()
-			tm, ext, err := ixdtf.Parse(input, true)
-			if err != nil {
-				t.Fatalf("Parse(%q, true) unexpected error: %v", input, err)
-			}
-
-			formatted, err := ixdtf.Format(tm, ext)
-			if err != nil {
-				t.Fatalf("Format after Parse(%q) unexpected error: %v", input, err)
-			}
-			if formatted != input {
-				t.Fatalf("round trip = %q, want %q", formatted, input)
-			}
-
-			if _, _, reparseErr := ixdtf.Parse(formatted, true); reparseErr != nil {
-				t.Fatalf("re-Parse(%q, true) unexpected error: %v", formatted, reparseErr)
-			}
-			if validateErr := ixdtf.Validate(formatted, true); validateErr != nil {
-				t.Fatalf("Validate(%q, true) unexpected error: %v", formatted, validateErr)
-			}
-		})
-	}
-}
-
-// TestParseErrorUnwrap verifies that ParseError supports errors.Is matching
-// against the package's sentinel errors via Unwrap.
-func TestParseErrorUnwrap(t *testing.T) {
-	t.Parallel()
-
-	_, _, err := ixdtf.Parse("2025-06-01T12:00:00+09:00[America/New_York]", true)
-	if !errors.Is(err, ixdtf.ErrTimezoneOffsetMismatch) {
-		t.Fatalf("expected errors.Is(err, ErrTimezoneOffsetMismatch), got %v", err)
-	}
-
-	_, _, err = ixdtf.Parse("2022-07-08T00:14:07Z[Asia/Tokyo][Europe/Paris]", true)
-	if !errors.Is(err, ixdtf.ErrInvalidSuffix) {
-		t.Fatalf("expected errors.Is(err, ErrInvalidSuffix), got %v", err)
-	}
-}
-
-func TestFormatNano(t *testing.T) {
-	t.Parallel()
-	tokyo, paris, cet := getTestTimezones()
-
-	tests := []struct {
-		name    string
-		tm      time.Time
-		ext     *ixdtf.IXDTFExtensions
-		want    string
-		wantErr string
-	}{
-		{
-			name: "basic with nanoseconds",
-			tm:   time.Date(2025, 1, 2, 3, 4, 5, 123456789, time.UTC),
-			ext:  ixdtf.NewIXDTFExtensions(nil),
-			want: "2025-01-02T03:04:05.123456789Z",
-		},
-		{
-			name: "with timezone and nanoseconds",
-			tm:   time.Date(2025, 2, 3, 4, 5, 6, 789000000, tokyo),
-			ext:  ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{Location: tokyo}),
-			want: "2025-02-03T04:05:06.789+09:00[Asia/Tokyo]",
-		},
-		{
-			name: "with tags and nanoseconds",
-			tm:   time.Date(2025, 3, 4, 5, 6, 7, 500000000, time.UTC),
-			ext: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{
-				Tags: map[string]string{"u-ca": "gregory"},
-			}),
-			want: "2025-03-04T05:06:07.5Z[u-ca=gregory]",
-		},
-		{
-			name: "with timezone, tags and nanoseconds",
-			tm:   time.Date(2025, 6, 7, 8, 9, 10, 250000000, paris),
-			ext: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{
-				Location: paris,
-				Tags:     map[string]string{"u-ca": "gregory"},
-				Critical: map[string]bool{"u-ca": true},
-			}),
-			want: "2025-06-07T08:09:10.25+01:00[Europe/Paris][!u-ca=gregory]",
-		},
-		{
-			name: "CET timezone with nanoseconds",
-			tm:   time.Date(2025, 12, 25, 15, 30, 45, 999000000, cet),
-			ext:  ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{Location: cet}),
-			want: "2025-12-25T15:30:45.999+01:00[CET]",
-		},
-		{
-			name: "multiple tags with different prefixes and nanoseconds",
-			tm:   time.Date(2025, 3, 4, 5, 6, 7, 500000000, time.UTC),
-			ext: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{
-				Tags: map[string]string{
-					"t-calendar": "japanese",
-					"u-ca":       "gregory",
-				},
-			}),
-			want: "2025-03-04T05:06:07.5Z[t-calendar=japanese][u-ca=gregory]",
-		},
-		{
-			name: "zero nanoseconds",
-			tm:   time.Date(2025, 1, 1, 0, 0, 0, 0, time.UTC),
-			ext:  ixdtf.NewIXDTFExtensions(nil),
-			want: "2025-01-01T00:00:00Z",
-		},
-		{
-			name: "invalid critical extension - empty value",
-			tm:   time.Date(2025, 1, 1, 0, 0, 0, 0, time.UTC),
-			ext: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{
-				Tags:     map[string]string{"u-ca": ""},
-				Critical: map[string]bool{"u-ca": true},
-			}),
-			wantErr: "critical extension cannot be processed",
-		},
-		{
-			name: "invalid critical calendar value",
-			tm:   time.Date(2025, 1, 1, 0, 0, 0, 0, time.UTC),
-			ext: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{
-				Tags:     map[string]string{"u-ca": "hoge"},
-				Critical: map[string]bool{"u-ca": true},
-			}),
-			wantErr: "invalid calendar tag identifier",
-		},
-		{
-			name: "missing critical extension",
-			tm:   time.Date(2025, 1, 1, 0, 0, 0, 0, time.UTC),
-			ext: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{
-				Tags:     map[string]string{},
-				Critical: map[string]bool{"u-ca": true},
-			}),
-			wantErr: "critical extension cannot be processed",
-		},
-		{
-			name: "invalid tag key format",
-			tm:   time.Date(2025, 1, 1, 0, 0, 0, 0, time.UTC),
-			ext: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{
-				Tags: map[string]string{"INVALID-KEY": "value"},
-			}),
-			wantErr: "invalid extension format",
-		},
-	}
-
-	sort.Slice(tests, func(i, j int) bool { return tests[i].name < tests[j].name })
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			t.Parallel()
-			got, err := ixdtf.FormatNano(tc.tm, tc.ext)
-
-			if tc.wantErr != "" {
-				if err == nil {
-					t.Fatalf("FormatNano(%v, %+v) expected error containing %q, got nil", tc.tm, tc.ext, tc.wantErr)
-				}
-				if !strings.Contains(err.Error(), tc.wantErr) {
-					t.Fatalf(
-						"FormatNano(%v, %+v) error = %q, want error containing %q",
-						tc.tm,
-						tc.ext,
-						err.Error(),
-						tc.wantErr,
-					)
-				}
-				return
-			}
-
-			if err != nil {
-				t.Fatalf("FormatNano(%v, %+v) unexpected error: %v", tc.tm, tc.ext, err)
-			}
-
-			if got != tc.want {
-				t.Errorf("FormatNano(%v, %+v) = %q, want %q", tc.tm, tc.ext, got, tc.want)
-			}
-		})
-	}
-}
-
-func TestParse(t *testing.T) {
-	t.Parallel()
-	tokyo, paris, cet := getTestTimezones()
-
-	tests := []struct {
-		name     string
-		input    string
-		strict   bool
-		wantTime time.Time
-		wantExt  *ixdtf.IXDTFExtensions
-		wantErr  string
-	}{
-		{
-			name:     "basic RFC3339",
-			input:    "2025-01-02T03:04:05Z",
-			strict:   false,
-			wantTime: time.Date(2025, 1, 2, 3, 4, 5, 0, time.UTC),
-			wantExt:  ixdtf.NewIXDTFExtensions(nil),
-		},
-		{
-			name:     "RFC3339 with nanoseconds",
-			input:    "2025-01-02T03:04:05.123456789Z",
-			strict:   false,
-			wantTime: time.Date(2025, 1, 2, 3, 4, 5, 123456789, time.UTC),
-			wantExt:  ixdtf.NewIXDTFExtensions(nil),
-		},
-		{
-			// RFC 9557 Section 3.4 (Figure 2): Z is an unknown local offset, so
-			// pairing it with a timezone is consistent even in strict mode.
-			name:     "Z with IANA timezone is consistent in strict mode",
-			input:    "2022-07-08T00:14:07Z[Europe/London]",
-			strict:   true,
-			wantTime: time.Date(2022, 7, 8, 0, 14, 7, 0, time.UTC),
-			wantExt: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{
-				Location: time.FixedZone("Europe/London", 0),
-			}),
-		},
-		{
-			name:     "with timezone",
-			input:    "2025-02-03T04:05:06+09:00[Asia/Tokyo]",
-			strict:   false,
-			wantTime: time.Date(2025, 2, 3, 4, 5, 6, 0, tokyo),
-			wantExt:  ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{Location: tokyo}),
-		},
-		{
-			name:     "with tags",
-			input:    "2025-03-04T05:06:07Z[u-ca=gregory]",
-			strict:   false,
-			wantTime: time.Date(2025, 3, 4, 5, 6, 7, 0, time.UTC),
-			wantExt: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{
-				Tags: map[string]string{"u-ca": "gregory"},
-			}),
-		},
-		{
-			name:     "with critical tag",
-			input:    "2025-03-04T05:06:07Z[!u-ca=gregory]",
-			strict:   false,
-			wantTime: time.Date(2025, 3, 4, 5, 6, 7, 0, time.UTC),
-			wantExt: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{
-				Tags:     map[string]string{"u-ca": "gregory"},
-				Critical: map[string]bool{"u-ca": true},
-			}),
-		},
-		{
-			name:     "with invalid calendar tag (non-critical)",
-			input:    "2025-03-04T05:06:07Z[u-ca=hoge]",
-			strict:   false,
-			wantTime: time.Date(2025, 3, 4, 5, 6, 7, 0, time.UTC),
-			wantExt: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{
-				Tags: map[string]string{"u-ca": "hoge"},
-			}),
-		},
-		{
-			name:    "with invalid calendar tag (non-critical, strict)",
-			input:   "2025-03-04T05:06:07Z[u-ca=hoge]",
-			strict:  true,
-			wantErr: "invalid calendar tag identifier",
-		},
-		{
-			name:    "with invalid calendar tag (critical)",
-			input:   "2025-03-04T05:06:07Z[!u-ca=hoge]",
-			strict:  false,
-			wantErr: "invalid calendar tag identifier",
-		},
-		{
-			name:     "with timezone and tags",
-			input:    "2025-06-07T08:09:10+01:00[Europe/Paris][!u-ca=gregory]",
-			strict:   false,
-			wantTime: time.Date(2025, 6, 7, 8, 9, 10, 0, paris),
-			wantExt: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{
-				Location: paris,
-				Tags:     map[string]string{"u-ca": "gregory"},
-				Critical: map[string]bool{"u-ca": true},
-			}),
-		},
-		{
-			name:     "multiple tags with different prefixes",
-			input:    "2025-03-04T05:06:07Z[t-calendar=japanese][u-ca=gregory]",
-			strict:   false,
-			wantTime: time.Date(2025, 3, 4, 5, 6, 7, 0, time.UTC),
-			wantExt: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{
-				Tags: map[string]string{
-					"t-calendar": "japanese",
-					"u-ca":       "gregory",
-				},
-			}),
-		},
-		{
-			name:     "multiple tags with critical flag",
-			input:    "2025-03-04T05:06:07Z[!t-format=iso][u-ca=hebrew]",
-			strict:   false,
-			wantTime: time.Date(2025, 3, 4, 5, 6, 7, 0, time.UTC),
-			wantExt: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{
-				Tags: map[string]string{
-					"t-format": "iso",
-					"u-ca":     "hebrew",
-				},
-				Critical: map[string]bool{
-					"t-format": true,
-				},
-			}),
-		},
-		{
-			name:     "timezone with CET",
-			input:    "2025-12-25T15:30:45+01:00[CET]",
-			strict:   false,
-			wantTime: time.Date(2025, 12, 25, 15, 30, 45, 0, cet),
-			wantExt:  ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{Location: cet}),
-		},
-		{
-			name:     "invalid timezone",
-			input:    "2025-01-01T00:00:00Z[Foo/Bar]",
-			strict:   false,
-			wantTime: time.Date(2025, 1, 1, 0, 0, 0, 0, time.UTC),
-			wantExt: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{
-				Location: nil,
-				Tags:     map[string]string{},
-				Critical: map[string]bool{},
-			}),
-		},
-		{
-			name:    "invalid timezone in strict mode",
-			input:   "2025-01-01T00:00:00Z[Foo/Bar]",
-			strict:  true,
-			wantErr: "invalid timezone",
-		},
-		{
-			name:    "malformed suffix",
-			input:   "2025-01-01T00:00:00Z[unclosed",
-			strict:  false,
-			wantErr: "IXDTFE parsing time",
-		},
-		{
-			name:    "empty string",
-			input:   "",
-			strict:  false,
-			wantErr: "IXDTFE parsing time",
-		},
-		{
-			name:    "invalid RFC3339",
-			input:   "not-a-date",
-			strict:  false,
-			wantErr: "IXDTFE parsing time",
-		},
-		{
-			name:     "timezone offset mismatch in non-strict mode",
-			input:    "2025-06-01T12:00:00+09:00[America/New_York]",
-			strict:   false,
-			wantTime: time.Date(2025, 6, 1, 12, 0, 0, 0, time.FixedZone("+0900", 9*3600)),
-			wantExt: ixdtf.NewIXDTFExtensions(
-				&ixdtf.NewIXDTFExtensionsArgs{Location: time.FixedZone("America/New_York", -4*3600)},
-			),
-		},
-		{
-			name:    "timezone offset mismatch in strict mode",
-			input:   "2025-06-01T12:00:00+09:00[America/New_York]",
-			strict:  true,
-			wantErr: "timezone offset does not match",
-		},
-		{
-			name:     "timezone with Etc/GMT pattern in non-strict mode",
-			input:    "2025-01-01T00:00:00+05:00[Etc/GMT-5]",
-			strict:   false,
-			wantTime: time.Date(2025, 1, 1, 0, 0, 0, 0, time.FixedZone("Etc/GMT-5", 5*3600)),
-			wantExt: ixdtf.NewIXDTFExtensions(
-				&ixdtf.NewIXDTFExtensionsArgs{Location: time.FixedZone("Etc/GMT-5", 5*3600)},
-			),
-		},
-		{
-			// Etc/GMT+3 is UTC-03:00 (POSIX-inverted sign); a concrete +05:00
-			// conflicts with it, and the critical flag forces the error even
-			// in non-strict mode (RFC 9557 Section 3.4).
-			name:    "critical Etc/GMT zone with mismatching offset errors in non-strict",
-			input:   "2022-07-08T00:14:07+05:00[!Etc/GMT+3]",
-			strict:  false,
-			wantErr: "timezone offset does not match",
-		},
-		{
-			name:    "Etc/GMT zone with mismatching offset errors in strict mode",
-			input:   "2022-07-08T00:14:07+05:00[Etc/GMT+3]",
-			strict:  true,
-			wantErr: "timezone offset does not match",
-		},
-		{
-			name:     "timezone with Etc/GMT pattern in strict mode",
-			input:    "2025-01-01T00:00:00+05:00[Etc/GMT-5]",
-			strict:   true,
-			wantTime: time.Date(2025, 1, 1, 0, 0, 0, 0, time.FixedZone("Etc/GMT-5", 5*3600)),
-			wantExt: ixdtf.NewIXDTFExtensions(
-				&ixdtf.NewIXDTFExtensionsArgs{Location: time.FixedZone("Etc/GMT-5", 5*3600)},
-			),
-		},
-		{
-			name:    "invalid suffix key format in parsing",
-			input:   "2025-01-01T00:00:00Z[INVALID-KEY=value]",
-			strict:  false,
-			wantErr: "IXDTFE parsing time",
-		},
-		{
-			// "[t-invalid]" has no "=" so it can only be a time-zone
-			// annotation, which the RFC 9557 Section 4.1 grammar
-			// ("suffix = [time-zone] *suffix-tag") forbids after a suffix tag.
-			name:    "timezone-shaped element after tag - non-strict mode",
-			input:   "2025-01-01T00:00:00Z[!u-ca=gregory][t-invalid]",
-			strict:  false,
-			wantErr: "invalid IXDTF suffix format",
-		},
-		{
-			name:    "timezone-shaped element after tag - strict mode",
-			input:   "2025-01-01T00:00:00Z[!u-ca=gregory][t-invalid]",
-			strict:  true,
-			wantErr: "invalid IXDTF suffix format",
-		},
-		{
-			name:    "timezone annotation after suffix tag rejected in strict",
-			input:   "2025-03-04T05:06:07Z[u-ca=hebrew][Asia/Tokyo]",
-			strict:  true,
-			wantErr: "invalid IXDTF suffix format",
-		},
-		{
-			name:    "timezone annotation after suffix tag rejected in non-strict",
-			input:   "2025-03-04T05:06:07Z[u-ca=hebrew][Asia/Tokyo]",
-			strict:  false,
-			wantErr: "invalid IXDTF suffix format",
-		},
-		{
-			// Even when the first zone is unknown and ignored by a non-strict
-			// parse, a second time-zone annotation stays a grammar violation;
-			// it must not be silently applied in place of the first.
-			name:    "second timezone annotation after ignored unknown zone rejected",
-			input:   "2022-07-08T00:14:07Z[Foo/Bar][Asia/Tokyo]",
-			strict:  false,
-			wantErr: "invalid IXDTF suffix format",
-		},
-		{
-			name:    "suffix with private extension",
-			input:   "2025-01-01T00:00:00Z[x-private=test]",
-			strict:  false,
-			wantErr: "private extension cannot be processed",
-		},
-		{
-			name:    "suffix with experimental extension",
-			input:   "2025-01-01T00:00:00Z[_experiment=test]",
-			strict:  false,
-			wantErr: "experimental extension cannot be processed",
-		},
-		{
-			name:    "critical private extension",
-			input:   "2025-01-01T00:00:00Z[!x-private=test]",
-			strict:  false,
-			wantErr: "private extension cannot be processed",
-		},
-		{
-			name:    "critical experimental extension",
-			input:   "2025-01-01T00:00:00Z[!_experiment=test]",
-			strict:  false,
-			wantErr: "experimental extension cannot be processed",
-		},
-		{
-			// RFC 9557 Section 4.1 permits a critical "!" flag on a time zone;
-			// combined with Z (unknown offset) it is consistent (Figure 2).
-			name:     "critical timezone with Z is accepted",
-			input:    "2025-01-01T00:00:00Z[!Asia/Tokyo]",
-			strict:   false,
-			wantTime: time.Date(2025, 1, 1, 0, 0, 0, 0, time.UTC),
-			wantExt: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{
-				Location:         time.FixedZone("Asia/Tokyo", 9*3600),
-				CriticalLocation: true,
-			}),
-		},
-		{
-			// Critical + concrete offset that conflicts with the zone: an
-			// application MUST act on the inconsistency even in non-strict
-			// mode (RFC 9557 Section 3.4).
-			name:    "critical timezone inconsistent offset errors in non-strict",
-			input:   "2022-07-08T00:14:07+00:00[!Europe/London]",
-			strict:  false,
-			wantErr: "timezone offset does not match",
-		},
-		{
-			// "-00:00" is the other unknown-local-offset form (RFC 9557
-			// Section 2.2), so a critical zone pairs with it consistently.
-			name:     "critical timezone with -00:00 is accepted",
-			input:    "2022-07-08T00:14:07-00:00[!Europe/London]",
-			strict:   false,
-			wantTime: time.Date(2022, 7, 8, 0, 14, 7, 0, time.UTC),
-			wantExt: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{
-				Location:         time.FixedZone("Europe/London", 1*3600),
-				CriticalLocation: true,
-			}),
-		},
-		{
-			// A critical annotation MUST be processable (Section 3.3), so an
-			// unknown name errors even in non-strict mode.
-			name:    "critical unknown timezone errors in non-strict",
-			input:   "2025-01-01T00:00:00Z[!Foo/Bar]",
-			strict:  false,
-			wantErr: "invalid timezone",
-		},
-		{
-			// The grammar allows at most one time-zone annotation (RFC 9557
-			// Section 4.1); a second one would overwrite the zone and its
-			// critical flag, hiding a mandatory Section 3.4 inconsistency error.
-			name:    "second timezone annotation rejected in non-strict",
-			input:   "2022-07-08T00:14:07+00:00[!Europe/London][Asia/Tokyo]",
-			strict:  false,
-			wantErr: "invalid IXDTF suffix format",
-		},
-		{
-			name:    "second timezone annotation rejected in strict mode",
-			input:   "2025-01-01T00:00:00Z[Asia/Tokyo][Europe/Paris]",
-			strict:  true,
-			wantErr: "invalid IXDTF suffix format",
-		},
-		{
-			name:     "timezone numeric offset suffix",
-			input:    "2025-01-01T00:00:00Z[+09:00]",
-			strict:   false,
-			wantTime: time.Date(2025, 1, 1, 0, 0, 0, 0, time.UTC),
-			wantExt: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{
-				Location: time.FixedZone("+09:00", 9*3600),
-				Tags:     map[string]string{},
-				Critical: map[string]bool{},
-			}),
-		},
-		{
-			// RFC 9557 Section 4.1 permits a critical numeric-offset
-			// annotation; a matching offset is fully processable
-			// (Section 3.3) and consistent (Section 3.4).
-			name:     "critical matching numeric offset is accepted",
-			input:    "2025-01-01T00:00:00+09:00[!+09:00]",
-			strict:   false,
-			wantTime: time.Date(2025, 1, 1, 0, 0, 0, 0, time.FixedZone("+09:00", 9*3600)),
-			wantExt: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{
-				Location:         time.FixedZone("+09:00", 9*3600),
-				CriticalLocation: true,
-			}),
-		},
-		{
-			name:    "critical mismatching numeric offset errors in non-strict",
-			input:   "2025-01-01T00:00:00+09:00[!+05:30]",
-			strict:  false,
-			wantErr: "timezone offset does not match",
-		},
-		{
-			// An offset-derived FixedZone must not be resolved via the
-			// timezone database, so a matching offset annotation is valid in
-			// strict mode too.
-			name:     "matching numeric offset is accepted in strict mode",
-			input:    "2025-01-01T00:00:00+09:00[+09:00]",
-			strict:   true,
-			wantTime: time.Date(2025, 1, 1, 0, 0, 0, 0, time.FixedZone("+09:00", 9*3600)),
-			wantExt: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{
-				Location: time.FixedZone("+09:00", 9*3600),
-			}),
-		},
-		{
-			name:     "invalid numeric offset ignored in non-strict mode",
-			input:    "2025-01-01T00:00:00Z[+24:00]",
-			strict:   false,
-			wantTime: time.Date(2025, 1, 1, 0, 0, 0, 0, time.UTC),
-			wantExt:  ixdtf.NewIXDTFExtensions(nil),
-		},
-		{
-			name:    "invalid numeric offset rejected in strict mode",
-			input:   "2025-01-01T00:00:00Z[+24:00]",
-			strict:  true,
-			wantErr: "invalid timezone name",
-		},
-		{
-			name:     "duplicate tag first wins",
-			input:    "2025-01-01T00:00:00Z[key=one][key=two]",
-			strict:   false,
-			wantTime: time.Date(2025, 1, 1, 0, 0, 0, 0, time.UTC),
-			wantExt: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{Tags: map[string]string{
-				"key": "one",
-			}}),
-		},
-		{
-			// RFC 9557 Section 3.3 example: a duplicate suffix key where the
-			// first occurrence is critical MUST be treated as erroneous.
-			name:    "critical duplicate tag rejected (first critical)",
-			input:   "2022-07-08T00:14:07Z[!u-ca=chinese][u-ca=japanese]",
-			strict:  false,
-			wantErr: "critical extension cannot be processed",
-		},
-		{
-			// RFC 9557 Section 3.3 example: the critical flag on the second
-			// occurrence must not be silently discarded either.
-			name:    "critical duplicate tag rejected (second critical)",
-			input:   "2022-07-08T00:14:07Z[u-ca=chinese][!u-ca=japanese]",
-			strict:  false,
-			wantErr: "critical extension cannot be processed",
-		},
-		{
-			name:    "critical duplicate tag rejected in strict mode",
-			input:   "2022-07-08T00:14:07Z[!u-ca=chinese][u-ca=japanese]",
-			strict:  true,
-			wantErr: "critical extension cannot be processed",
-		},
-		{
-			// RFC 9557 Section 3.3 example "[!knort=blargel]": in strict mode
-			// this library is the recipient, understands only "u-ca", and MUST
-			// treat an unrecognized critical key as erroneous.
-			name:    "unknown critical suffix key rejected in strict mode",
-			input:   "2022-07-08T00:14:07Z[!knort=blargel]",
-			strict:  true,
-			wantErr: "critical extension cannot be processed",
-		},
-		{
-			// In non-strict mode processing of unrecognized critical keys is
-			// delegated to the caller via the Critical map.
-			name:     "unknown critical suffix key delegated in non-strict mode",
-			input:    "2022-07-08T00:14:07Z[!knort=blargel]",
-			strict:   false,
-			wantTime: time.Date(2022, 7, 8, 0, 14, 7, 0, time.UTC),
-			wantExt: ixdtf.NewIXDTFExtensions(&ixdtf.NewIXDTFExtensionsArgs{
-				Tags:     map[string]string{"knort": "blargel"},
-				Critical: map[string]bool{"knort": true},
-			}),
-		},
-	}
-
-	sort.Slice(tests, func(i, j int) bool { return tests[i].name < tests[j].name })
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			t.Parallel()
-			gotTime, gotExt, err := ixdtf.Parse(tc.input, tc.strict)
-
-			if tc.wantErr != "" {
-				checkParseError(t, err, tc.input, tc.strict, tc.wantErr)
-				return
-			}
-
-			if err != nil {
-				t.Fatalf("Parse(%q, %v) unexpected error: %v", tc.input, tc.strict, err)
-			}
-
-			compareParseResults(t, gotTime, gotExt, tc.input, tc.strict, tc.wantTime, tc.wantExt)
-		})
-	}
-}
-
-func TestValidateUnicodeCalendarIdentifiers(t *testing.T) {
-	t.Parallel()
-
-	validIdentifiers := []string{
-		"buddhist",
-		"chinese",
-		"coptic",
-		"dangi",
-		"ethioaa",
-		"ethiopic-amete-alem",
-		"ethiopic",
-		"gregory",
-		"gregorian",
-		"hebrew",
-		"indian",
-		"islamic",
-		"islamic-umalqura",
-		"islamic-tbla",
-		"islamic-civil",
-		"islamic-rgsa",
-		"iso8601",
-		"japanese",
-		"persian",
-		"roc",
-		"islamicc",
-	}
-
-	for _, identifier := range validIdentifiers {
-		t.Run(identifier, func(t *testing.T) {
-			t.Parallel()
-			input := "2025-03-04T05:06:07Z[!u-ca=" + identifier + "]"
-			if err := ixdtf.Validate(input, false); err != nil {
-				t.Fatalf("Validate(%q) unexpected error: %v", input, err)
-			}
-		})
-	}
-}
-
-func TestValidate(t *testing.T) {
-	t.Parallel()
-	tests := []struct {
-		name    string
-		input   string
-		strict  bool
-		wantErr string
-	}{
-		{
-			name:   "valid RFC3339",
-			input:  "2025-01-02T03:04:05Z",
-			strict: false,
-		},
-		{
-			name:   "valid RFC3339 with nanoseconds",
-			input:  "2025-01-02T03:04:05.123456789Z",
-			strict: false,
-		},
-		{
-			name:   "valid with timezone",
-			input:  "2025-02-03T04:05:06Z[Asia/Tokyo]",
-			strict: false,
-		},
-		{
-			name:   "valid with offset and timezone",
-			input:  "2025-02-03T04:05:06+09:00[Asia/Tokyo]",
-			strict: false,
-		},
-		{
-			name:   "valid with tags",
-			input:  "2025-03-04T05:06:07Z[u-ca=gregory]",
-			strict: false,
-		},
-		{
-			name:   "valid with critical tag",
-			input:  "2025-03-04T05:06:07Z[!u-ca=gregory]",
-			strict: false,
-		},
-		{
-			// Validate must accept a critical zone exactly like Parse does
-			// (RFC 9557 Section 3.4, Figure 2: Z carries no local offset).
-			name:   "valid with critical timezone",
-			input:  "2022-07-08T00:14:07Z[!Europe/London]",
-			strict: false,
-		},
-		{
-			name:   "valid with critical numeric offset",
-			input:  "2025-01-01T00:00:00+09:00[!+09:00]",
-			strict: false,
-		},
-		{
-			name:   "valid with invalid calendar tag (non-critical)",
-			input:  "2025-03-04T05:06:07Z[u-ca=hoge]",
-			strict: false,
-		},
-		{
-			name:    "invalid calendar tag (non-critical, strict)",
-			input:   "2025-03-04T05:06:07Z[u-ca=hoge]",
-			strict:  true,
-			wantErr: "IXDTFE parsing time \"2025-03-04T05:06:07Z[u-ca=hoge]\" as \"2006-01-02T15:04:05Z07:00*([time-zone-name][tags])\": invalid calendar tag identifier",
-		},
-		{
-			name:    "invalid calendar tag (critical)",
-			input:   "2025-03-04T05:06:07Z[!u-ca=hoge]",
-			strict:  false,
-			wantErr: "IXDTFE parsing time \"2025-03-04T05:06:07Z[!u-ca=hoge]\" as \"2006-01-02T15:04:05Z07:00*([time-zone-name][tags])\": invalid calendar tag identifier",
-		},
-		{
-			name:   "valid with timezone and tags",
-			input:  "2025-06-07T08:09:10+01:00[Europe/Paris][!u-ca=gregory]",
-			strict: false,
-		},
-		{
-			name:   "invalid timezone name in non-strict mode",
-			input:  "2025-01-01T00:00:00Z[No/SuchZone]",
-			strict: false,
-		},
-		{
-			name:    "invalid timezone name in strict mode",
-			input:   "2025-01-01T00:00:00Z[No/SuchZone]",
-			strict:  true,
-			wantErr: "IXDTFE parsing time \"2025-01-01T00:00:00Z[No/SuchZone]\" as \"2006-01-02T15:04:05Z07:00*([time-zone-name][tags])\": invalid timezone name",
-		},
-		{
-			name:   "invalid extension format in non-strict mode",
-			input:  "2025-01-01T00:00:00Z[invalid]",
-			strict: false,
-		},
-		{
-			name:    "invalid extension format in strict mode",
-			input:   "2025-01-01T00:00:00Z[invalid]",
-			strict:  true,
-			wantErr: "IXDTFE parsing time \"2025-01-01T00:00:00Z[invalid]\" as \"2006-01-02T15:04:05Z07:00*([time-zone-name][tags])\": invalid timezone name",
-		},
-		{
-			name:    "private extension",
-			input:   "2025-01-01T00:00:00Z[x-demo=yes]",
-			strict:  false,
-			wantErr: "IXDTFE parsing time \"2025-01-01T00:00:00Z[x-demo=yes]\" as \"2006-01-02T15:04:05Z07:00*([time-zone-name][tags])\": private extension cannot be processed",
-		},
-		{
-			name:    "experimental extension",
-			input:   "2025-01-01T00:00:00Z[_test=value]",
-			strict:  false,
-			wantErr: "IXDTFE parsing time \"2025-01-01T00:00:00Z[_test=value]\" as \"2006-01-02T15:04:05Z07:00*([time-zone-name][tags])\": experimental extension cannot be processed",
-		},
-		{
-			name:    "empty string",
-			input:   "",
-			strict:  false,
-			wantErr: "IXDTFE parsing time \"\" as \"2006-01-02T15:04:05Z07:00\": empty datetime string",
-		},
-		{
-			name:    "invalid RFC3339",
-			input:   "not-a-date",
-			strict:  false,
-			wantErr: "IXDTFE parsing time \"not-a-date\" as \"2006-01-02T15:04:05Z07:00\": invalid portion: parsing time \"not-a-date\" as \"2006-01-02T15:04:05Z07:00\": cannot parse \"not-a-date\" as \"2006\"",
-		},
-		{
-			name:    "malformed suffix",
-			input:   "2025-01-01T00:00:00Z[unclosed",
-			strict:  false,
-			wantErr: "IXDTFE parsing time \"2025-01-01T00:00:00Z[unclosed\" as \"2006-01-02T15:04:05Z07:00*([time-zone-name][tags])\": invalid IXDTF suffix format",
-		},
-		{
-			name:   "timezone offset mismatch in non-strict mode",
-			input:  "2025-06-01T12:00:00+09:00[America/New_York]",
-			strict: false,
-		},
-		{
-			name:    "timezone offset mismatch in strict mode",
-			input:   "2025-06-01T12:00:00+09:00[America/New_York]",
-			strict:  true,
-			wantErr: "IXDTFE parsing time \"2025-06-01T12:00:00+09:00[America/New_York]\" as \"2006-01-02T15:04:05.999999999Z07:00*([time-zone-name][tags])\": timezone offset does not match the specified timezone",
-		},
-		{
-			name:   "timezone content with u- prefix - non-strict mode",
-			input:  "2025-01-01T00:00:00Z[u-invalid-timezone]",
-			strict: false,
-		},
-		{
-			name:   "timezone content with x- prefix - non-strict mode",
-			input:  "2025-01-01T00:00:00Z[x-invalid-timezone]",
-			strict: false,
-		},
-		{
-			name:    "timezone content with u- prefix - strict mode",
-			input:   "2025-01-01T00:00:00Z[u-invalid-timezone]",
-			strict:  true,
-			wantErr: "IXDTFE parsing time \"2025-01-01T00:00:00Z[u-invalid-timezone]\" as \"2006-01-02T15:04:05Z07:00*([time-zone-name][tags])\": invalid timezone name",
-		},
-		{
-			name:    "timezone content with x- prefix - strict mode",
-			input:   "2025-01-01T00:00:00Z[x-invalid-timezone]",
-			strict:  true,
-			wantErr: "IXDTFE parsing time \"2025-01-01T00:00:00Z[x-invalid-timezone]\" as \"2006-01-02T15:04:05Z07:00*([time-zone-name][tags])\": invalid timezone name",
-		},
-		{
-			name:    "suffix with invalid characters in key",
-			input:   "2025-01-01T00:00:00Z[invalid@key=value]",
-			strict:  false,
-			wantErr: "IXDTFE parsing time \"2025-01-01T00:00:00Z[invalid@key=value]\" as \"2006-01-02T15:04:05Z07:00*([time-zone-name][tags])\": invalid extension format",
-		},
-		{
-			name:    "suffix with invalid characters in value range",
-			input:   "2025-01-01T00:00:00Z[key=invalid@value]",
-			strict:  false,
-			wantErr: "IXDTFE parsing time \"2025-01-01T00:00:00Z[key=invalid@value]\" as \"2006-01-02T15:04:05Z07:00*([time-zone-name][tags])\": invalid extension format",
-		},
-		{
-			name:    "critical extension with invalid value range",
-			input:   "2025-01-01T00:00:00Z[!key=invalid@value]",
-			strict:  false,
-			wantErr: "IXDTFE parsing time \"2025-01-01T00:00:00Z[!key=invalid@value]\" as \"2006-01-02T15:04:05Z07:00*([time-zone-name][tags])\": invalid extension format",
-		},
-		{
-			name:    "critical indicator without timezone content",
-			input:   "2025-01-01T00:00:00Z[!]",
-			strict:  false,
-			wantErr: "IXDTFE parsing time \"2025-01-01T00:00:00Z[!]\" as \"2006-01-02T15:04:05Z07:00*([time-zone-name][tags])\": invalid IXDTF suffix format",
-		},
-		{
-			name:   "invalid numeric offset ignored in non-strict mode",
-			input:  "2025-01-01T00:00:00Z[+24:00]",
-			strict: false,
-		},
-		{
-			name:    "invalid numeric offset rejected in strict mode",
-			input:   "2025-01-01T00:00:00Z[+24:00]",
-			strict:  true,
-			wantErr: "IXDTFE parsing time \"2025-01-01T00:00:00Z[+24:00]\" as \"2006-01-02T15:04:05Z07:00*([time-zone-name][tags])\": invalid timezone name",
-		},
-		{
-			name:    "invalid timezone syntax rejected by ABNF",
-			input:   "2025-01-01T00:00:00Z[invalid@zone]",
-			strict:  false,
-			wantErr: "IXDTFE parsing time \"2025-01-01T00:00:00Z[invalid@zone]\" as \"2006-01-02T15:04:05Z07:00*([time-zone-name][tags])\": invalid extension format",
-		},
-		{
-			name:   "valid suffix with hyphenated key",
-			input:  "2025-01-01T00:00:00Z[key-with-hyphens=value]",
-			strict: false,
-		},
-		{
-			name:   "valid suffix with hyphenated value",
-			input:  "2025-01-01T00:00:00Z[key=value-with-hyphens]",
-			strict: false,
-		},
-		{
-			name:   "valid suffix with numeric value",
-			input:  "2025-01-01T00:00:00Z[key=123]",
-			strict: false,
-		},
-		{
-			name:   "valid suffix with mixed alphanumeric",
-			input:  "2025-01-01T00:00:00Z[key=abc123def]",
-			strict: false,
-		},
-		{
-			name:    "empty suffix key",
-			input:   "2025-01-01T00:00:00Z[=value]",
-			strict:  false,
-			wantErr: "IXDTFE parsing time \"2025-01-01T00:00:00Z[=value]\" as \"2006-01-02T15:04:05Z07:00*([time-zone-name][tags])\": invalid extension format",
-		},
-		{
-			name:    "empty suffix value",
-			input:   "2025-01-01T00:00:00Z[key=]",
-			strict:  false,
-			wantErr: "IXDTFE parsing time \"2025-01-01T00:00:00Z[key=]\" as \"2006-01-02T15:04:05Z07:00*([time-zone-name][tags])\": invalid extension format",
-		},
-		{
-			name:    "empty timezone brackets",
-			input:   "2025-01-01T00:00:00Z[]",
-			strict:  false,
-			wantErr: "IXDTFE parsing time \"2025-01-01T00:00:00Z[]\" as \"2006-01-02T15:04:05Z07:00*([time-zone-name][tags])\": invalid IXDTF suffix format",
-		},
-		{
-			name:    "suffix key starting with number",
-			input:   "2025-01-01T00:00:00Z[123key=value]",
-			strict:  false,
-			wantErr: "IXDTFE parsing time \"2025-01-01T00:00:00Z[123key=value]\" as \"2006-01-02T15:04:05Z07:00*([time-zone-name][tags])\": invalid extension format",
-		},
-		{
-			name:    "suffix key with spaces",
-			input:   "2025-01-01T00:00:00Z[key with spaces=value]",
-			strict:  false,
-			wantErr: "IXDTFE parsing time \"2025-01-01T00:00:00Z[key with spaces=value]\" as \"2006-01-02T15:04:05Z07:00*([time-zone-name][tags])\": invalid extension format",
-		},
-		{
-			name:    "suffix value with underscore",
-			input:   "2025-01-01T00:00:00Z[key=val_ue]",
-			strict:  false,
-			wantErr: "IXDTFE parsing time \"2025-01-01T00:00:00Z[key=val_ue]\" as \"2006-01-02T15:04:05Z07:00*([time-zone-name][tags])\": invalid extension format",
-		},
-	}
-
-	sort.Slice(tests, func(i, j int) bool { return tests[i].name < tests[j].name })
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			t.Parallel()
-			err := ixdtf.Validate(tc.input, tc.strict)
-			if (err != nil && err.Error() != tc.wantErr) || (err == nil && tc.wantErr != "") {
-				t.Fatalf("Validate(%q, %v) error = %v, wantErr %v", tc.input, tc.strict, err, tc.wantErr)
-			}
-		})
-	}
-}
-
-func checkParseError(t *testing.T, err error, input string, strict bool, wantErr string) {
-	if err == nil {
-		t.Fatalf("Parse(%q, %v) expected error containing %q, got nil", input, strict, wantErr)
-	}
-	if !strings.Contains(err.Error(), wantErr) {
-		t.Fatalf(
-			"Parse(%q, %v) error = %q, want error containing %q",
-			input,
-			strict,
-			err.Error(),
-			wantErr,
-		)
-	}
-}
-
-func compareParseResults(
-	t *testing.T,
-	gotTime time.Time,
-	gotExt *ixdtf.IXDTFExtensions,
-	input string,
-	strict bool,
-	wantTime time.Time,
-	wantExt *ixdtf.IXDTFExtensions,
-) {
-	if !gotTime.Equal(wantTime) {
-		t.Errorf("Parse(%q, %v) time = %v, want %v", input, strict, gotTime, wantTime)
-	}
-
-	if !extensionsEqual(gotExt, wantExt) {
-		t.Errorf("Parse(%q, %v) extensions = %+v, want %+v", input, strict, gotExt, wantExt)
-	}
-}
-
-func extensionsEqual(a, b *ixdtf.IXDTFExtensions) bool {
-	if a == nil && b == nil {
-		return true
-	}
-	if a == nil || b == nil {
-		return false
-	}
-
-	if (a.Location == nil) != (b.Location == nil) {
-		return false
-	}
-	if a.Location != nil && b.Location != nil {
-		if a.Location.String() != b.Location.String() {
-			return false
-		}
-	}
-
-	if a.CriticalLocation != b.CriticalLocation {
-		return false
-	}
-
-	if len(a.Tags) != len(b.Tags) {
-		return false
-	}
-	for k, v := range a.Tags {
-		if b.Tags[k] != v {
-			return false
-		}
-	}
-
-	if len(a.Critical) != len(b.Critical) {
-		return false
-	}
-	for k, v := range a.Critical {
-		if b.Critical[k] != v {
-			return false
-		}
-	}
-
-	return true
-}
-
-func getTestTimezones() (*time.Location, *time.Location, *time.Location) {
-	return time.FixedZone("Asia/Tokyo", 9*3600),
-		time.FixedZone("Europe/Paris", 1*3600),
-		time.FixedZone("CET", 1*3600)
-}
-
-// TestParseUnknownLocalOffsetAppliesTimezone verifies RFC 9557 Section 3.4
-// (Figure 2): when the RFC 3339 part uses an unknown local offset ("Z" or
-// "-00:00"), the time-zone annotation is applied to resolve local time, and
-// this is never treated as an inconsistency — in either strict or non-strict
-// mode. See https://github.com/8beeeaaat/ixdtf/issues/22.
-func TestParseUnknownLocalOffsetAppliesTimezone(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name       string
-		input      string
-		wantOffset int // seconds east of UTC after applying the timezone
-		wantHour   int // local wall-clock hour after applying the timezone
-	}{
-		{
-			name:       "Z with London in summer applies BST (+01:00)",
-			input:      "2022-07-08T00:14:07Z[Europe/London]",
-			wantOffset: 1 * 3600,
-			wantHour:   1,
-		},
-		{
-			name:       "Z with London in winter applies GMT (+00:00)",
-			input:      "2022-01-08T00:14:07Z[Europe/London]",
-			wantOffset: 0,
-			wantHour:   0,
-		},
-		{
-			name:       "negative-zero offset with London applies BST",
-			input:      "2022-07-08T00:14:07-00:00[Europe/London]",
-			wantOffset: 1 * 3600,
-			wantHour:   1,
-		},
-		{
-			name:       "Z with New York in summer applies EDT (-04:00)",
-			input:      "2022-07-08T12:00:00Z[America/New_York]",
-			wantOffset: -4 * 3600,
-			wantHour:   8,
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			t.Parallel()
-			for _, strict := range []bool{false, true} {
-				got, ext, err := ixdtf.Parse(tc.input, strict)
-				if err != nil {
-					t.Fatalf("Parse(%q, %v) unexpected error: %v", tc.input, strict, err)
-				}
-				if _, off := got.Zone(); off != tc.wantOffset {
-					t.Errorf("Parse(%q, %v) offset = %d, want %d", tc.input, strict, off, tc.wantOffset)
-				}
-				if got.Hour() != tc.wantHour {
-					t.Errorf("Parse(%q, %v) local hour = %d, want %d", tc.input, strict, got.Hour(), tc.wantHour)
-				}
-				if ext.Location == nil {
-					t.Errorf("Parse(%q, %v) expected location to be set", tc.input, strict)
-				}
-			}
-		})
-	}
-}
-
-// TestParseCriticalTimezone verifies RFC 9557 handling of a critical "!" flag
-// on a time-zone annotation (Section 4.1 grammar; Section 3.3/3.4 semantics).
-// See the follow-up to https://github.com/8beeeaaat/ixdtf/issues/22.
-func TestParseCriticalTimezone(t *testing.T) {
-	t.Parallel()
-
-	t.Run("Z with critical zone is consistent in both modes", func(t *testing.T) {
-		t.Parallel()
-		const input = "2022-07-08T00:14:07Z[!Europe/London]" // Figure 2: no inconsistency
-		for _, strict := range []bool{false, true} {
-			got, ext, err := ixdtf.Parse(input, strict)
-			if err != nil {
-				t.Fatalf("Parse(%q, %v) unexpected error: %v", input, strict, err)
-			}
-			if !ext.CriticalLocation {
-				t.Errorf("Parse(%q, %v) expected CriticalLocation to be true", input, strict)
-			}
-			if _, off := got.Zone(); off != 1*3600 { // London is BST (+01:00) in July
-				t.Errorf("Parse(%q, %v) offset = %d, want %d", input, strict, off, 1*3600)
-			}
-		}
-	})
-
-	t.Run("critical zone round-trips with the ! flag", func(t *testing.T) {
-		t.Parallel()
-		const input = "2025-01-01T00:00:00+09:00[!Asia/Tokyo]"
-		got, ext, err := ixdtf.Parse(input, true)
-		if err != nil {
-			t.Fatalf("Parse(%q) unexpected error: %v", input, err)
-		}
-		formatted, err := ixdtf.Format(got, ext)
-		if err != nil {
-			t.Fatalf("Format unexpected error: %v", err)
-		}
-		if formatted != input {
-			t.Errorf("round trip failed: got %q, want %q", formatted, input)
-		}
-	})
-
-	t.Run("critical inconsistency errors in non-strict mode", func(t *testing.T) {
-		t.Parallel()
-		// +00:00 asserts a zero offset that conflicts with London's +01:00 in
-		// July; the critical flag forces the error even in non-strict mode.
-		const input = "2022-07-08T00:14:07+00:00[!Europe/London]"
-		_, _, err := ixdtf.Parse(input, false)
-		checkParseError(t, err, input, false, "timezone offset does not match")
-	})
-}
-
-// TestParseNumericZeroOffsetStillInconsistent verifies the contrast with Z: a
-// concrete "+00:00" asserts a zero offset, so it conflicts with a timezone
-// whose offset differs at that instant (RFC 9557 Section 3.4, Figure 1). This
-// behavior must remain unchanged by the Z fix.
-func TestParseNumericZeroOffsetStillInconsistent(t *testing.T) {
-	t.Parallel()
-
-	const input = "2022-07-08T00:14:07+00:00[Europe/London]" // London is BST (+01:00) in July
-
-	// strict mode: the inconsistency is reported as an error.
-	if _, _, err := ixdtf.Parse(input, true); err == nil {
-		t.Fatalf("Parse(%q, true) expected inconsistency error, got nil", input)
-	}
-
-	// non-strict mode: the original +00:00 offset is preserved (timezone not applied).
-	got, _, err := ixdtf.Parse(input, false)
-	if err != nil {
-		t.Fatalf("Parse(%q, false) unexpected error: %v", input, err)
-	}
-	if _, off := got.Zone(); off != 0 {
-		t.Errorf("Parse(%q, false) offset = %d, want 0 (original preserved)", input, off)
-	}
-}
